requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pypdf

langchain
langchain-community
langchain-groq

pinecone langchain-pinecone
langchain-huggingface

sentence-transformers
huggingface-hub

pydantic
requests
tqdm
